<마이페이지> + 클릭 이벤트 추가 + 레이아웃 변경 + 옷장, 북마크 '마이페이지' 하위 페키지로 넣음
