+++ app/src/main/java/org/techtown/sns_project/Board/Upload/closet/closet_shoes_adapter.java
@@ -2,7 +2,6 @@ package org.techtown.sns_project.Board.Upload.closet;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,7 +13,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 
-import org.techtown.sns_project.Closet.Closet_info;
+import org.techtown.sns_project.fragment.profile.Closet.Closet_info;
 import org.techtown.sns_project.R;
 
 import java.util.ArrayList;
