+++ app/src/main/java/org/techtown/sns_project/fragment/profile/Closet/AllFragment.java
@@ -1,4 +1,4 @@
-package org.techtown.sns_project.Closet;
+package org.techtown.sns_project.fragment.profile.Closet;
 
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -24,13 +24,11 @@ import com.google.firebase.firestore.QueryDocumentSnapshot;
 
 import org.techtown.sns_project.Camera.Activity_codi;
 import org.techtown.sns_project.R;
-import org.techtown.sns_project.qr.ProductInfo;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 
-public class shoesFragment extends Fragment {
+public class AllFragment extends Fragment {
 
     static FirebaseAuth firebaseAuth = FirebaseAuth.getInstance();
     static FirebaseUser firebaseUser = firebaseAuth.getCurrentUser();
@@ -49,17 +47,17 @@ public class shoesFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        View v = inflater.inflate(R.layout.activity_closet_shoes, container, false);
+        View v = inflater.inflate(R.layout.activity_closet_all, container, false);
         Closet_adapter = new ClosetAdapter();
         //recyclerview
-        recyclerView = v.findViewById(R.id.shoes_Recyclerview);
+        recyclerView = v.findViewById(R.id.all_Recyclerview);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new GridLayoutManager(getActivity(),2));
         recyclerView.setAdapter(Closet_adapter);
 
         //파베에서 옷 정보 가져와서 어뎁터에 전달
         scatter();
-        ClosetMainActivity.whatFragment("shoes");
+        ClosetMainActivity.whatFragment("all");
 
         //클릭시 삭제 or 코디
         Closet_adapter.setOnItemClickListener(new ClosetAdapter.OnItemClickListener() {
@@ -112,7 +110,9 @@ public class shoesFragment extends Fragment {
             }
         });
 
+
         return v;
+
     }
 
     //파베에서 옷 정보 가져와서 어뎁터에 전달
@@ -120,14 +120,14 @@ public class shoesFragment extends Fragment {
 
         Closet_adapter.list.clear();
         Closet_adapter.notifyDataSetChanged();
-        db.collection("users").document(firebaseUser.getUid()).collection("신발").get().
+        db.collection("users").document(firebaseUser.getUid()).collection("아우터").get().
                 addOnCompleteListener(task -> {
                     if (task.isSuccessful()) {
-                        //데이터 중복 방지
 
                         for (QueryDocumentSnapshot document : task.getResult()) {
 
                             List = (HashMap<String, Object>) document.getData();
+
                             Closet_info data = new Closet_info(
                                     (String)List.get("name"),
                                     (String)List.get("brand"),
@@ -138,10 +138,77 @@ public class shoesFragment extends Fragment {
                             Closet_adapter.addItem(data);
                         }
 
-                        Closet_adapter.notifyDataSetChanged();
+                    } else {
+                        Log.d(TAG, "Error getting documents: ", task.getException());
+                    }
+                });
+        db.collection("users").document(firebaseUser.getUid()).collection("상의").get().
+                addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+
+                            List = (HashMap<String, Object>) document.getData();
+
+                            Closet_info data = new Closet_info(
+                                    (String)List.get("name"),
+                                    (String)List.get("brand"),
+                                    (String)List.get("clothes_type"),
+                                    (String)List.get("img_url"),
+                                    (String) List.get("url"));
+
+                            Closet_adapter.addItem(data);
+                        }
+
+                    } else {
+                        Log.d(TAG, "Error getting documents: ", task.getException());
+                    }
+                });
+        db.collection("users").document(firebaseUser.getUid()).collection("바지").get().
+                addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+
+                            List = (HashMap<String, Object>) document.getData();
+
+                            Closet_info data = new Closet_info(
+                                    (String)List.get("name"),
+                                    (String)List.get("brand"),
+                                    (String)List.get("clothes_type"),
+                                    (String)List.get("img_url"),
+                                    (String) List.get("url"));
+
+                            Closet_adapter.addItem(data);
+                        }
+
+                    } else {
+                        Log.d(TAG, "Error getting documents: ", task.getException());
+                    }
+                });
+        db.collection("users").document(firebaseUser.getUid()).collection("신발").get().
+                addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+
+                            List = (HashMap<String, Object>) document.getData();
+
+                            Closet_info data = new Closet_info(
+                                    (String)List.get("name"),
+                                    (String)List.get("brand"),
+                                    (String)List.get("clothes_type"),
+                                    (String)List.get("img_url"),
+                                    (String) List.get("url"));
+
+                            Closet_adapter.addItem(data);
+                        }
+
                     } else {
                         Log.d(TAG, "Error getting documents: ", task.getException());
                     }
                 });
     }
+
+
 }
