+++ app/src/main/java/org/techtown/sns_project/fragment/profile/Closet/shoesFragment.java
@@ -1,4 +1,4 @@
-package org.techtown.sns_project.Closet;
+package org.techtown.sns_project.fragment.profile.Closet;
 
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -19,20 +19,16 @@ import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 
 import org.techtown.sns_project.Camera.Activity_codi;
-import org.techtown.sns_project.Enterprise.QR.EnterpriseQRListActivity;
 import org.techtown.sns_project.R;
-import org.techtown.sns_project.qr.ProductInfo;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 
-public class bottomFragment extends Fragment {
+public class shoesFragment extends Fragment {
 
     static FirebaseAuth firebaseAuth = FirebaseAuth.getInstance();
     static FirebaseUser firebaseUser = firebaseAuth.getCurrentUser();
@@ -51,17 +47,17 @@ public class bottomFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        View v = inflater.inflate(R.layout.activity_closet_bottom, container, false);
+        View v = inflater.inflate(R.layout.activity_closet_shoes, container, false);
         Closet_adapter = new ClosetAdapter();
         //recyclerview
-        recyclerView = v.findViewById(R.id.bottom_Recyclerview);
+        recyclerView = v.findViewById(R.id.shoes_Recyclerview);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new GridLayoutManager(getActivity(),2));
         recyclerView.setAdapter(Closet_adapter);
 
         //파베에서 옷 정보 가져와서 어뎁터에 전달
         scatter();
-        ClosetMainActivity.whatFragment("bottom");
+        ClosetMainActivity.whatFragment("shoes");
 
         //클릭시 삭제 or 코디
         Closet_adapter.setOnItemClickListener(new ClosetAdapter.OnItemClickListener() {
@@ -122,7 +118,7 @@ public class bottomFragment extends Fragment {
 
         Closet_adapter.list.clear();
         Closet_adapter.notifyDataSetChanged();
-        db.collection("users").document(firebaseUser.getUid()).collection("바지").get().
+        db.collection("users").document(firebaseUser.getUid()).collection("신발").get().
                 addOnCompleteListener(task -> {
                     if (task.isSuccessful()) {
                         //데이터 중복 방지
