+++ app/src/main/java/org/techtown/sns_project/fragment/profile/ProfileFragment.java
@@ -38,10 +38,10 @@ import com.gun0912.tedpermission.TedPermission;
 import com.theartofdev.edmodo.cropper.CropImage;
 
 import org.techtown.sns_project.Board.BoardPostClickEvent;
-import org.techtown.sns_project.Closet.ClosetMainActivity;
-import org.techtown.sns_project.Closet.Closet_info;
+import org.techtown.sns_project.fragment.profile.Closet.ClosetMainActivity;
 import org.techtown.sns_project.R;
 import org.techtown.sns_project.fragment.DataFormat;
+import org.techtown.sns_project.fragment.profile.Bookmark.bookmark;
 import org.techtown.sns_project.qr.ProductInfo;
 
 import java.util.ArrayList;
